Added Exceptions and messages.properties file for error describing.

+++ src/main/java/com/savaleks/angularspring/rest/UserRegistrationRestController.java
@@ -1,6 +1,7 @@
 package com.savaleks.angularspring.rest;
 
 import com.savaleks.angularspring.dto.UserDTO;
+import com.savaleks.angularspring.exception.CustomErrorType;
 import com.savaleks.angularspring.repo.UserJpaRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,6 +11,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -28,11 +30,20 @@ public class UserRegistrationRestController {
     @GetMapping("/")
     public ResponseEntity<List<UserDTO>> listAllUsers(){
         List<UserDTO> users = userJpaRepository.findAll();
+        if (users.isEmpty()){
+            return new ResponseEntity<List<UserDTO>>(HttpStatus.NO_CONTENT);
+        }
         return new ResponseEntity<List<UserDTO>>(users, HttpStatus.OK);
     }
 
     @PostMapping(value = "/", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<UserDTO> createUser(@RequestBody final UserDTO user){
+    public ResponseEntity<UserDTO> createUser(@Valid @RequestBody final UserDTO user){
+        LOGGER.info("Creating user: {}", user);
+        if (userJpaRepository.findByName(user.getName()) != null){
+            return new ResponseEntity<UserDTO>(new CustomErrorType(
+                    "Unable to create new user. A user " + user.getName() + " already exists"
+            ), HttpStatus.CONFLICT);
+        }
         userJpaRepository.save(user);
         return new ResponseEntity<UserDTO>(user, HttpStatus.CREATED);
     }
@@ -40,6 +51,10 @@ public class UserRegistrationRestController {
     @GetMapping("/{id}")
     public ResponseEntity<UserDTO> getUserById(@PathVariable("id") final Long id){
         UserDTO user = userJpaRepository.findById(id).orElse(null);
+        if (user == null){
+            return new ResponseEntity<UserDTO>(
+                    new CustomErrorType("User with id " + id + " not found"), HttpStatus.NOT_FOUND);
+        }
         return new ResponseEntity<UserDTO>(user, HttpStatus.OK);
     }
 
@@ -48,6 +63,13 @@ public class UserRegistrationRestController {
 
         // retrieve user data based on id and set it to user object of type UserDTO
         UserDTO currentUser = userJpaRepository.findById(id).orElse(null);
+
+        if (currentUser == null){
+            return new ResponseEntity<UserDTO>(
+                    new CustomErrorType("Unable to update. User with id " + id + " not found."),
+                    HttpStatus.NOT_FOUND
+            );
+        }
 
         // update user object data with user object data
         currentUser.setName(user.getName());
@@ -63,7 +85,16 @@ public class UserRegistrationRestController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<UserDTO> deleteUser(@PathVariable("id") final Long id){
+        // find user in database by id, if not exists throw exception
+        UserDTO user = userJpaRepository.findById(id).orElse(null);
+        if (user == null){
+            return new ResponseEntity<UserDTO>(
+                    new CustomErrorType("Unable to delete. User with id " + id + " not found."),
+                    HttpStatus.NOT_FOUND);
+        }
         userJpaRepository.deleteById(id);
-        return new ResponseEntity<UserDTO>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity<UserDTO>(
+                new CustomErrorType("Deleted user with id " + id + "."),
+                HttpStatus.NO_CONTENT);
     }
 }
